+++ R/pull.R
@@ -15,6 +15,12 @@
 pull.RPolarsDataFrame <- function(.data, var, ...) {
   check_polars_data(.data)
   var <- tidyselect_named_arg(.data, rlang::enquo(var))
+  if (length(var) > 1) {
+    rlang::abort(
+      paste0("`pull` can only extract one column. You tried to extract ", length(var), ".")
+    )
+  }
+
   out <- add_tidypolars_class(.data)
   if (inherits(.data, "RPolarsLazyFrame")) {
     out <- out$collect()